requirements.txt
// Python 3.9.21
geoopt==0.5.0
geotorch==0.3.0
matplotlib==3.9.4
networkx==3.2.1
numpy==1.26.3
ogb==1.3.6
pandas==2.2.3
pip==25.0
scikit-learn==1.6.1
scipy==1.13.1
torch==2.2.0+cu118
torch_cluster==1.6.3+pt22cu118
torch-geometric==2.7.0
torch_scatter==2.1.2+pt22cu118
torch_sparse==0.6.18+pt22cu118
torchaudio==2.2.0+cu118
torchvision==0.17.0+cu118
tqdm==4.67.1
umap-learn==0.5.7

# torch==2.2.0 torchvision==0.17.0 torchaudio==2.2.0 --index-url https://download.pytorch.org/whl/cu118
# torchvision==0.17.0 --index-url https://download.pytorch.org/whl/cu118
# torchaudio==2.2.0 --index-url https://download.pytorch.org/whl/cu118
# torch-scatter torch-sparse torch-cluster -f https://data.pyg.org/whl/torch-2.2.0+cu118.html
# torch-sparse -f https://data.pyg.org/whl/torch-2.2.0+cu118.html
# torch-cluster -f https://data.pyg.org/whl/torch-2.2.0+cu118.html
# git+https://github.com/pyg-team/pytorch_geometric.git